requirements.txt
numpy==1.22.4
tweepy==4.10.0
python-dotenv==0.20.0
